changed Navbar component and used component in page.js
